[R] enable R compile
[R] Enable R build for windows and linux
