+++ amalgamation/dmlc-minimum0.cc
@@ -1,0 +1,14 @@
+/*!
+ * Copyright 2015 by Contributors.
+ * \brief Mininum DMLC library Amalgamation, used for easy plugin of dmlc lib.
+ *  Normally this is not needed.
+ */
+#include "../dmlc-core/src/io/line_split.cc"
+#include "../dmlc-core/src/io/recordio_split.cc"
+#include "../dmlc-core/src/io/input_split_base.cc"
+#include "../dmlc-core/src/io/local_filesys.cc"
+#include "../dmlc-core/src/data.cc"
+#include "../dmlc-core/src/io.cc"
+#include "../dmlc-core/src/recordio.cc"
+
+
